Refactor references to Byron data in /blockchain/
Many files that should be neutral such as the database.js, cron.js, etc
were directly referencing Byron-format blockchain data in the
/blockchain/ folder. This will now have a common interface for data that
is needed by these entities with the exception of Txs.

Txs will be handled later, but for now the temporary plan is to just get
Shelley working with UTXOs and later we can figure out how to handle Tx
generalization with certificates/etc.

This probably warrants further testing. Only basic block-pulling and
storing was tested. The api/txs endpoint was not, nor rollbacks, etc,
etc. This commit is more for design feedback to make sure this is a
solid approach and to build upon it.