Discriminate UTXO vs Account outputs
This is done at the parsing stage for convenience.
It can be obtained by the address stored within, but is easiest to do
here when we already have the wasm structure loaded.
Every output will need this anyway as we need to treat utxo vs account
differently during DB storage.